client: include http status code with the error
This PR includes now the http status code as part of the returned error.
Users of this client can type assert the error value to get more
information about the underling error.

We also remove the error to open an issues to planetscale-go repository
as that is not very helpful for both the user and for us.