Update logic to be agnostic for the private key type
Go has the crypto.PrivateKey interface to represent a type agnostic
private key.

The code here is updated to remove the *rsa.PrivateKey assumption and
instead to be agnostic for the specific type of key. In the few places
we need to detect the type, we can do so appropriately (mostly in
serialization logic).