Move to the full access host instead of the remote addr
The remote addr is only a base address and not the full address. This
means clients still build their own hostnames which is what we want to
move away from.

This changes the internal API and we'll be updating all the dependencies
then to switch away from RemoteAddr to this address.