passwords: support listing all passwords
The API wasn't able to list all passwords. However both in the UI and in
the CLI we want to be able to show all existing passwords.

This PR changes the `passwords.List` command to show all existing
passwords for a given database. It also supports filtering according to
a branch, if given.