Switch to new certificates endpoint for generation
We're moving anything based on certificates to a new endpoint, so this
adds support for that endpoint. The API is changing slightly here to
build the tls.Certificate object so that we can also more easily
retrieve the object for a GET request where the private key is not
known (since tls.Certificate requires it).